+++ lib/advanced_value_notifier.dart
@@ -4,5 +4,6 @@ export 'src/notifiers/simple_notifier.dart';
 export 'src/builder/simple_builder.dart';
 export 'src/listener/simple_listener.dart';
 
-export 'src/notifiers/transformed_notifier.dart';
+export 'src/notifiers/transformer_notifier.dart';
 export 'src/builder/transformer_builder.dart';
+export 'src/listener/transformer_listener.dart';