+++ CHANGELOG.md
@@ -1,6 +1,9 @@
+## 0.0.3
+* Added `TransformerHistoryValueListenableListener`
+
 ## 0.0.2
-* Added transformer notifier and builder
+* Added `TransformerHistoryValueNotifier` and `HistoryValueListenableBuilder`
 
 ## 0.0.1
 
-* Added HistoryValueNotifier HistoryValueListenableListener and HistoryValueListenableBuilder
+* Added `HistoryValueNotifier`, `HistoryValueListenableListener` and `HistoryValueListenableBuilder`